Fix typo in RegistrySync that sends and receives RESTORE for APPLY. (#227)
Note: This requires a version bump and backward compatibility code to avoid
breaking the client by the server not sending expected data. This won't break
the server as the server is an authority, and won't require either the server
or client to update as is with the compatibility code in place.